+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 
+2.0.5 -> 2.0.10:
+* Editable Assembly test and checklist
+* 'Load microcode' has been integrated into the 'File name'
+
 2.0.2 -> 2.0.5:
 * Update to bootstrap 4
 * Internal code updated: now input files and results are independent